finally figured out weird bug with allocating memory for the string, yay for not really knowing wtf I'm doing in c

+++ Entry.h
@@ -1,6 +1,6 @@
 struct Entry {
     char *key;
-    unsigned long value;
+    int value;
     struct Entry *next;
 };
